geo_audit_agent/requirements.txt
python-dotenv
google-genai
langgraph>=0.2.0
pydantic
tenacity
plotly
scikit-learn
mlflow
sentence-transformers
chromadb
extruct
beautifulsoup4
requests
pandas
numpy
